% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functons.R
\name{days_before_date}
\alias{days_before_date}
\title{days_before_date}
\usage{
days_before_date(input_date, days_before)
}
\arguments{
\item{x}{The input parameter.}
}
\value{
The pattern.
}
\description{
This function sets a pattern.
}
