% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functons.R
\name{get_train_matrix}
\alias{get_train_matrix}
\title{get_train_matrix}
\usage{
get_train_matrix(
  num_lags,
  laglag,
  pred_length,
  timee,
  data,
  removed_names,
  data_pattern
)
}
\arguments{
\item{x}{The input parameter.}
}
\value{
The pattern.
}
\description{
This function sets a pattern.
}
