% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functons.R
\name{Predicting}
\alias{Predicting}
\title{Predicting}
\usage{
Predicting(
  train_pattern,
  nn,
  pred_length,
  num_lags,
  datasum,
  position,
  model_list
)
}
\arguments{
\item{x}{The input parameter.}
}
\value{
The pattern.
}
\description{
This function sets a pattern.
}
