% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functons.R
\name{set_pattern}
\alias{set_pattern}
\title{set_pattern}
\usage{
set_pattern(
  data_raw,
  data_changed,
  data_ori,
  data_minus,
  threshold = 0.05,
  cut_time = "2021-01-01"
)
}
\arguments{
\item{x}{The input parameter.}
}
\value{
The pattern.
}
\description{
This function sets a pattern.
}
