% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functons.R
\name{Train_ML}
\alias{Train_ML}
\title{Train_ML}
\usage{
Train_ML(
  train_patterns_list,
  path,
  MLE_meth_name,
  svm_type = "eps-regression",
  XG_eta = 0.3,
  XG_max_depth = 6,
  XG_eval_metric = "rmse",
  XG_num_rounds = 300
)
}
\arguments{
\item{x}{The input parameter.}
}
\value{
The pattern.
}
\description{
This function sets a pattern.
}
